+++ frontend/app/src/entities/events/ui/node-events/node-event-title.tsx
@@ -1,34 +1,10 @@
-import { ACCOUNT_OBJECT, PROPOSED_CHANGES_OBJECT } from "@/config/constants";
-import { QSP } from "@/config/qsp";
 import { NODE_EVENTS_MAPPING } from "@/entities/events/ui/node-events/constants";
 import { NodeLabel } from "@/entities/nodes/object/ui/node-label";
 import { schemaKindLabelState } from "@/entities/schema/stores/schemaKindLabel.atom";
 import type { NodeMutatedEvent } from "@/shared/api/graphql/generated/graphql";
-import { constructPath } from "@/shared/api/rest/fetch";
 import { Link } from "@/shared/components/ui/link";
 import { useAtomValue } from "jotai";
-
-export const getLink = ({ kind, id, branch }: { kind: string; id: string; branch: string }) => {
-  if (kind === PROPOSED_CHANGES_OBJECT) {
-    return constructPath(`/proposed-changes/${id}`);
-  }
-
-  if (kind === ACCOUNT_OBJECT) {
-    return constructPath("/role-management", [
-      {
-        name: QSP.BRANCH,
-        value: branch,
-      },
-    ]);
-  }
-
-  return constructPath(`/objects/${kind}/${id}`, [
-    {
-      name: QSP.BRANCH,
-      value: branch,
-    },
-  ]);
-};
+import { getLink } from "./utils";
 
 export const NodeEventTitle = ({
   event,
